authorization added and tested with valid void and invalid tokens. PUT updated and fixed. New functionality in JWTManager to support jwt in the app

+++ src/Controller/UserController.php
@@ -10,6 +10,7 @@ namespace App\Controller;
 
 use App\Errors;
 use App\Entity\User;
+use App\JWTManager;
 use Symfony\Bundle\FrameworkBundle\Controller\AbstractController;
 use Symfony\Component\HttpFoundation\JsonResponse;
 use Symfony\Component\HttpFoundation\Response;
@@ -26,11 +27,16 @@ class UserController extends AbstractController
     const USER_API_PATH = '/api/v1/users';
 
     /**
+     * @param Request $request
      * @return JsonResponse
      * @Route(path="",name="getc",methods={Request::METHOD_GET})
      */
-    public function getCUser(): JsonResponse
+    public function getCUser(Request $request): JsonResponse
     {
+        $authoritation = $this->authorization($request);
+        if ($authoritation !== true) {
+            return $authoritation;
+        }
         /** @var Errors $error */
         $error = new Errors();
         $em = $this->getDoctrine()->getManager();
@@ -43,10 +49,15 @@ class UserController extends AbstractController
     /**
      * @Route(path="/{id}",name="get",methods={Request::METHOD_GET})
      * @param int id
+     * @param Request $request
      * @return JsonResponse
      */
-    public function getOneUser(int $id): JsonResponse
+    public function getOneUser(int $id, Request $request): JsonResponse
     {
+        $authoritation = $this->authorization($request);
+        if ($authoritation !== true) {
+            return $authoritation;
+        }
         /** @var Errors $error */
         $error = new Errors();
         $em = $this->getDoctrine()->getManager();
@@ -61,42 +72,46 @@ class UserController extends AbstractController
      * @param Request $request
      * @return JsonResponse
      */
-    public function postUser(Request $request):JsonResponse
+    public function postUser(Request $request): JsonResponse
     {
+        $authoritation = $this->authorization($request);
+        if ($authoritation !== true) {
+            return $authoritation;
+        }
         /** @var Errors $error */
         $error = new Errors();
         $em = $this->getDoctrine()->getManager();
         $datosPeticion = $request->getContent();
         $datos = json_decode($datosPeticion, true);
-        if (!array_key_exists('username', $datos)||!array_key_exists('email',$datos)||!array_key_exists('password',$datos)) {
+        if (!array_key_exists('username', $datos) || !array_key_exists('email', $datos) || !array_key_exists('password', $datos)) {
             return $error->error422();
         }
-        if ($em->getRepository(User::class)->findBy(['username'=>$datos['username']])||
-        $em->getRepository(User::class)->findBy(['email'=>$datos['email']])){
+        if ($em->getRepository(User::class)->findBy(['username' => $datos['username']]) ||
+            $em->getRepository(User::class)->findBy(['email' => $datos['email']])) {
             return $error->error400();
         }
         $user = new User(
             $datos['username'],
             $datos['email'],
             $datos['password'],
-            $datos['enabled']??true,
-            $datos['isAdmin']??false
+            $datos['enabled'] ?? true,
+            $datos['isAdmin'] ?? false
         );
         $em->persist($user);
         $em->flush();
-        return new JsonResponse(['user'=>$user],Response::HTTP_CREATED);
+        return new JsonResponse(['user' => $user], Response::HTTP_CREATED);
     }
 
     /**
      * @Route(path="",name="options",methods={Request::METHOD_OPTIONS})
      * @return JsonResponse
      */
-    public function optionsUser():JsonResponse
+    public function optionsUser(): JsonResponse
     {
         return new JsonResponse(
             'Allow header',
             Response::HTTP_OK,
-            ['allow'=>Request::METHOD_GET.', '.Request::METHOD_POST]
+            ['allow' => Request::METHOD_GET . ', ' . Request::METHOD_POST]
         );
     }
 
@@ -105,27 +120,32 @@ class UserController extends AbstractController
      * @param int id
      * @return JsonResponse
      */
-    public function optionsOneUser(int $id):JsonResponse
+    public function optionsOneUser(int $id): JsonResponse
     {
         return new JsonResponse(
             'Allow header',
             Response::HTTP_OK,
-            ['allow'=>Request::METHOD_GET.', '.Request::METHOD_PUT.', '.Request::METHOD_DELETE]
+            ['allow' => Request::METHOD_GET . ', ' . Request::METHOD_PUT . ', ' . Request::METHOD_DELETE]
         );
     }
 
     /**
      * @Route(path="/{id}",name="delete",methods={Request::METHOD_DELETE})
      * @param int $id
+     * @param Request $request
      * @return JsonResponse
      */
-    public function deleteUser(int $id):JsonResponse
+    public function deleteUser(int $id, Request $request): JsonResponse
     {
+        $authoritation = $this->authorization($request);
+        if ($authoritation !== true) {
+            return $authoritation;
+        }
         /** @var Errors $error */
         $error = new Errors();
         $em = $this->getDoctrine()->getManager();
         $user = $em->getRepository(User::class)->find($id);
-        if ($user===null){
+        if ($user === null) {
             return $error->error404();
         }
         $em->remove($user);
@@ -142,8 +162,12 @@ class UserController extends AbstractController
      * @param int $id
      * @return JsonResponse
      */
-    public function putUser(Request $request,int $id):JsonResponse
+    public function putUser(Request $request, int $id): JsonResponse
     {
+        $authoritation = $this->authorization($request);
+        if ($authoritation !== true) {
+            return $authoritation;
+        }
         /** @var Errors $error */
         $error = new Errors();
         $em = $this->getDoctrine()->getManager();
@@ -151,31 +175,51 @@ class UserController extends AbstractController
         $datos = json_decode($datosPeticion, true);
         /** @var User $user */
         $user = $em->getRepository(User::class)->find($id);
-        if ($user===null){
+        if ($user === null) {
             return $error->error404();
         }
-        if ($em->getRepository(User::class)->findBy(['username'=>$datos['username']])||
-            $em->getRepository(User::class)->findBy(['email'=>$datos['email']])){
-            return $error->error400();
+        if (isset($datos['username'])) {
+            if ($em->getRepository(User::class)->findBy(['username' => $datos['username']])) {
+                return $error->error400();
+            }
         }
-        if (isset($datos['username'])){
+        if (isset($datos['email'])) {
+            if ($em->getRepository(User::class)->findBy(['email' => $datos['email']])) {
+                return $error->error400();
+            }
+        }
+        if (isset($datos['username'])) {
             $user->setUsername($datos['username']);
         }
-        if (isset($datos['email'])){
+        if (isset($datos['email'])) {
             $user->setEmail($datos['email']);
         }
-        if (isset($datos['password'])){
+        if (isset($datos['password'])) {
             $user->setPassword($datos['password']);
         }
-        if (isset($datos['enabled'])){
+        if (isset($datos['enabled'])) {
             $user->setEnabled($datos['enabled']);
         }
-        if (isset($datos['isAdmin'])){
+        if (isset($datos['isAdmin'])) {
             $user->setIsAdmin($datos['isAdmin']);
         }
         $em->persist($user);
         $em->flush();
-        return new JsonResponse($user,209);
+        return new JsonResponse($user, 209);
     }
 
+    private function authorization(Request $request)
+    {
+        /** @var Errors $error */
+        $error = new Errors();
+        $jwtManager = new JWTManager();
+        $token = $request->headers->get('token');
+        if ($token === null || !$jwtManager->isValidToken($token)) {
+            return $error->error401();
+        }
+        if (!$jwtManager->isAdminUser($token)) {
+            return $error->error403();
+        }
+        return true;
+    }
 }
